# Conflicts:
#	common/lib/config_definition.py

+++ backend/workers/cleanup_tempfiles.py
@@ -3,13 +3,14 @@ Delete old datasets
 """
 import shutil
 import re
-
+import json
+from datetime import datetime
 from pathlib import Path
 
 from common.config_manager import config
 from backend.lib.worker import BasicWorker
 from common.lib.dataset import DataSet
-from common.lib.exceptions import WorkerInterruptedException
+from common.lib.exceptions import WorkerInterruptedException, DataSetException
 
 
 class TempFileCleaner(BasicWorker):
@@ -25,16 +26,32 @@ class TempFileCleaner(BasicWorker):
     type = "clean-temp-files"
     max_workers = 1
 
+    ensure_job = {"remote_id": "localhost", "interval": 10800}
+
+    # Use tracking file to delay deletion of files that may still be in use
+    tracking_file = config.get('PATH_DATA').joinpath(".temp_file_cleaner")
+    days_to_keep = 7
+
     def work(self):
         """
         Go through result files, and for each one check if it should still
         exist
         :return:
         """
+        # Load tracking file
+        if not self.tracking_file.exists():
+            tracked_files = {}
+        else:
+            tracked_files = json.loads(self.tracking_file.read_text())
 
         result_files = Path(config.get('PATH_DATA')).glob("*")
         for file in result_files:
+            if file.stem.startswith("."):
+                # skip hidden files
+                continue
+
             if self.interrupted:
+                self.tracking_file.write_text(json.dumps(tracked_files))
                 raise WorkerInterruptedException("Interrupted while cleaning up orphaned result files")
 
             # the key of the dataset files belong to can be extracted from the
@@ -50,18 +67,31 @@ class TempFileCleaner(BasicWorker):
 
             try:
                 dataset = DataSet(key=key, db=self.db)
-            except TypeError:
+            except DataSetException:
                 # the dataset has been deleted since, but the result file still
                 # exists - should be safe to clean up
-                self.log.info("No matching dataset with key %s for file %s, deleting file" % (key, str(file)))
-                if file.is_dir():
-                    shutil.rmtree(file)
-                else:
-                    try:
-                        file.unlink()
-                    except FileNotFoundError:
-                        # the file has been deleted since
-                        pass
+                if file.name not in tracked_files:
+                    self.log.info(f"No matching dataset with key {key} for file {file}; marking for deletion")
+                    tracked_files[file.name] = datetime.now().timestamp() + (self.days_to_keep * 86400)
+                elif tracked_files[file.name] < datetime.now().timestamp():
+                    self.log.info(f"File {file} marked for deletion since {datetime.fromtimestamp(tracked_files[file.name]).strftime('%Y-%m-%d %H:%M:%S')}, deleting file")
+                    if file.is_dir():
+                        try:
+                            shutil.rmtree(file)
+                        except PermissionError:
+                            self.log.info(f"Folder {file} does not belong to a dataset but cannot be deleted (no "
+                                          f"permissions), skipping")
+
+                    else:
+                        try:
+                            file.unlink()
+                        except FileNotFoundError:
+                            # the file has been deleted since
+                            pass
+
+                    # Remove from tracking
+                    del tracked_files[file.name]
+
                 continue
 
             if file.is_dir() and "-staging" in file.stem and dataset.is_finished():
@@ -69,6 +99,14 @@ class TempFileCleaner(BasicWorker):
                 # if the dataset is finished, the staging area should have been
                 # compressed into a zip file, or deleted, so this is also safe
                 # to clean up
-                self.log.debug("Dataset %s is finished, but staging area remains at %s, deleting folder" % (
-                dataset.key, str(file)))
-                shutil.rmtree(file)
+                if file.name not in tracked_files:
+                    self.log.info("Dataset %s is finished, but staging area remains at %s, marking for deletion" % (dataset.key, str(file)))
+                    tracked_files[file.name] = datetime.now().timestamp() + (self.days_to_keep * 86400)
+                elif tracked_files[file.name] < datetime.now().timestamp():
+                    self.log.info("Dataset %s is finished, but staging area remains at %s, deleting folder" % (dataset.key, str(file)))
+                    shutil.rmtree(file)
+
+        # Update tracked files
+        self.tracking_file.write_text(json.dumps(tracked_files))
+
+        self.job.finish()
